Fix root layout
